requirements.txt
editdistance==0.8.1
lxml==5.2.2
matplotlib==3.8.4
numpy==2.3.3
PyYAML==6.0.2
torch==2.5.1
tqdm==4.67.1